Encapsulate and document log-odds fusion scorers
